Rewritten printForWebgraphviz procedure to spot easily parallel edges
Add chcecks that permits while loading the graph from a DIMAC challenge file to load two edges a->b b->a with different costs.

Add a procedure that permits to genereate random graphs. Two approaches have been followed
1. using RMAT (the same used in the paper, https://www.boost.org/doc/libs/1_53_0/libs/graph_parallel/doc/html/scalable_rmat_generator.html#czf04);
2. Erdös-Renyi (https://www.boost.org/doc/libs/1_62_0/libs/graph/doc/erdos_renyi_generator.html)
Both are already provided by boost library. This pays back time spent replacing project datastructures with the one offered by boost library.
A word of advice : the paper was probably made to compute the MST of graph made by a single connected component. In order to force this after the random procedure that creates the graph two more passed are required to eliminate self loops and adding weights (attention: since undirected graph are targeted we have to make sure that the graph has both edges a->b b->a with the same cost).

Two naive ways to measure time taken for both algorithm to complete (<GPU, boruvka>, <CPU, mst>) have been implemented. Note that this has limited importance  since the two algorithm are completely different and therefore not comparable. In the paper the implementation comparison was between two different versio of boruvka algorithm.

For sake of completeness:
(a single instance of the problem, 4096 nodes, ~4096*4 edges, weights 1-1000)
- time taken by boruvka (gpu) :  ~628 ms
- time taken by  kruskal (cpu): ~228 sec = 3.8 min

note: DIMACS challenge smallests graph (NY) contains 264346 nodes ~(>2^18) and 733846 edges with weights that exceeds 8800 units(>2^13), which makes impossible to use the algorithm with the proposed bits configuration for VERTEX, WEIGHT in the segmented min scan.